Added logic to PlayHand, changed RollAllDice
The PlayHand method is now used within Playgame to allow the user to
keep dice or reroll them. Rewrote RollAllDice with for loop. Added
TODOs for things that need to be implemented. Currently no way to
control the scorecard entires. At the end of each turn the total is
added to the running total. Need to move the die into an array for
easier referencing. Need to implement a scorcard system.